backend/requirements.txt
djangorestframework==3.13.1
psycopg2==2.9.
djangorestframework-simplejwt==5.2.0
django-filter==22.1
django-cors-headers==3.13.0
pyyaml
uritemplate
drf-yasg
django-extensions
pygraphviz